增加了背景音樂，開車動畫，修正了死亡后鍵盤移動bug

+++ MagneticGangsta/Assets/Scripts/UI/LoadingPanel.cs
@@ -25,6 +25,6 @@ public class LoadingPanel : UIBase
     // Update is called once per frame
     void Update()
     {
-        m_loadingText.text = m_loadingTextStart + LoadSceneManager.Instance.LoadProgress * 100 + "%";
+        m_loadingText.text = m_loadingTextStart + (int)(LoadSceneManager.Instance.LoadProgress * 100) + "%";
     }
 }
